cloud_app/requirements-cloud.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
boto3>=1.34.0
python-dotenv>=1.0.0
